Update README.md
This is an application which uses the Spotify API to fetch the users Top Artists, and Top Tracks. It allows you to search Spotify to find songs and artists and has a built in player component at the bottom of each page. 
I have focused on making this application as accessible as possible by adding clear interactivity, as well as a colour and dark mode.